requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt   
pillow
psycopg2-binary
